Replace countries dependency with YAML file
The only reason phony_rails uses the countries gem is to be able to map country 2-letter codes to their numerical country_code. It doesn't really make sense to require an entire gem just for this functionality.

Furthermore, the countries gem uses up a lot of memory, as profiled by the derailed_benchmarks gem. For such simple functionality, it's better to put the info in a small YAML file rather than use a heavy gem when only a small portion of its code will be used.